requirements.txt
absl-py==1.4.0
astunparse==1.6.3
autopep8==1.6.0
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
conda==23.1.0
conda-content-trust==0.1.3
conda-package-handling==2.0.2
conda_package_streaming==0.7.0
cryptography==38.0.4
filelock==3.9.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
gviz-api==1.10.0
h5py==3.6.0
huggingface-hub==0.12.1
idna==3.4
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
more-itertools==9.1.0
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.8.6
packaging==23.0
pip==23.0.1
pluggy==1.0.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycosat==0.6.4
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
setuptools==65.6.3
six==1.15.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-profile==2.11.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.19.0
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.2
tensorflow-metal==0.5.0
termcolor==2.2.0
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
tqdm==4.64.1
transformers==4.26.1
typeguard==2.13.3
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wheel==0.37.1
wrapt==1.14.1
zstandard==0.18.0
